requirements.txt
pdfplumber>=0.10.0
Pillow>=10.0.0
PyPDF2>=3.0.0
lxml>=4.9.0
Jinja2>=3.1.0
python-dateutil>=2.8.0
ebooklib>=0.18.0
validators>=0.22.0
tqdm>=4.66.0
numpy>=1.20.0
